Adding filechecks to split_by_subset

+++ scripts/stats/cam/exevs_cam_nam_firewxnest_grid2obs_stats.sh
@@ -118,8 +118,16 @@ then
  mkdir -p $DATA/$OBSDIR/nam.${obday}
   cp -v $COMINobsproc/nam.${obday}/nam.t${obcyc}z.prepbufr.tm${tmnum} $DATA/$OBSDIR/nam.${obday}/nam.t${obcyc}z.prepbufr.tm${tmnum}
   split_by_subset $DATA/$OBSDIR/nam.${obday}/nam.t${obcyc}z.prepbufr.tm${tmnum}
-  cat ADPUPA ADPSFC MSONET SFCSHP >> prepbufr.tmp
-  mv prepbufr.tmp $DATA/$OBSDIR/nam.${obday}/nam.t${obcyc}z.prepbufr.tm${tmnum}
+  for subset in ADPUPA ADPSFC MSONET SFCSHP; do
+    if [ -e $subset ]; then
+     cat $subset >> prepbufr.tmp
+    fi
+  done
+  if [ -e prepbufr.tmp ]; then
+   mv prepbufr.tmp $DATA/$OBSDIR/nam.${obday}/nam.t${obcyc}z.prepbufr.tm${tmnum}
+  else
+   obfound=0
+  fi
 else
   echo "WARNING: File $COMINobsproc/${MODELNAME}.${obday}/${MODELNAME}.t${obcyc}z.prepbufr.tm${tmnum} is missing."
   if [ $SENDMAIL = "YES" ]; then
